requirements.txt
asgiref==3.4.1
caldav==0.8.0
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
django-filter==2.4.0
idna==3.2
lxml==4.6.3
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
uuid==1.30
vobject==0.9.6.1
